refactor: add vue-router
It was removed in https://github.com/LyonDataViz/occupapp/commit/e19055c838ba6117a22d2106decb8d4b955b5509
