+++ src/main.ts
@@ -1,11 +1,13 @@
 import Vue from 'vue'
 import App from './App.vue'
+import router from './router'
 import store from './store'
 import vuetify from './plugins/vuetify'
 
 Vue.config.productionTip = false
 
 new Vue({
+  router,
   store,
   vuetify,
   render: h => h(App)
